Parsing : make the underlying number type double
This needed a change in the parser, and since we may want to force
working with integers, a 'floor' operator has been added too.

Still end of chapter 9 on S-expr

+++ src/lval.h
@@ -9,7 +9,7 @@ enum { LERR_DIV_ZERO, LERR_BAD_OP, LERR_BAD_NUM };
 
 struct lval {
     int type;
-    long num;
+    double num;
 
     char* err;
     char* sym;
@@ -19,7 +19,7 @@ struct lval {
 };
 
 /* Create a new lval from a number */
-struct lval* lval_num(long x);
+struct lval* lval_num(double x);
 
 /* Create a new lval from an error */
 struct lval* lval_err(char* message);
